requirements.txt
alembic
arrow
awesome-slugify
beautifulsoup4
decorator
Mako
natsort
nose
nose-selecttests
gunicorn
httplib2
psycopg2
pyramid
pyramid-debugtoolbar
pyramid-mako
pyramid-tm
pyramid-celery
pyramid-simpleform
pyramid_storage
python-thumbnails
py-bcrypt
pymongo
PasteDeploy
requests
requests-futures
SQLAlchemy
sqlalchemy-utils
waitress
WebHelpers2
webtest
zope.sqlalchemy